RecStrip 1.0: ======== - NavProcessor (SD): B-Frames nach I-Frame werden ignoriert (-> testen!) - durchgängiger Support von sync-Verlusten (unvollständige/verschlüsselte Pakete, Header-Daten) - PacketSize wird auch für *.rec ermittelt - Teletext-Daten können ebenfalls entfernt werden - RebuildInf: TS-Analyse erfolgt nicht mehr am Anfang (wegen PVR5000-Header) - Support von Humax-Aufnahmen (1184 Byte Header in TS) - PMT/PAT-Generierung für Humax-Aufnahmen

+++ type.h
@@ -18,10 +18,10 @@ extern "C" {
 #endif
 
 #ifndef dword
-typedef unsigned long    dword;
+typedef unsigned int     dword;
 #endif
 #ifndef word
-typedef unsigned short    word;
+typedef unsigned short   word;
 #endif
 #ifndef byte
 typedef unsigned char    byte;
